Moves tests to align with package structure and makes warnings tests only check line for expected warning
