+++ tests/tools/test_warnings.py
@@ -1,0 +1,86 @@
+import os
+
+import numpy as np
+import pytest
+
+from parcels import (
+    AdvectionRK4_3D,
+    AdvectionRK45,
+    FieldSet,
+    FieldSetWarning,
+    KernelWarning,
+    ParticleSet,
+    ScipyParticle,
+)
+
+
+def test_fieldset_warnings():
+    # halo with inconsistent boundaries
+    lat = [0, 1, 5, 10]
+    lon = [0, 1, 5, 10]
+    u = [[1, 1, 1, 1] for _ in range(4)]
+    v = [[1, 1, 1, 1] for _ in range(4)]
+    fieldset = FieldSet.from_data(data={"U": u, "V": v}, dimensions={"lon": lon, "lat": lat}, transpose=True)
+    with pytest.warns(FieldSetWarning):
+        fieldset.add_periodic_halo(meridional=True, zonal=True)
+
+    # flipping warning
+    lat = [0, 1, 5, -5]
+    lon = [0, 1, 5, 10]
+    u = [[1, 1, 1, 1] for _ in range(4)]
+    v = [[1, 1, 1, 1] for _ in range(4)]
+    with pytest.warns(FieldSetWarning):
+        fieldset = FieldSet.from_data(data={"U": u, "V": v}, dimensions={"lon": lon, "lat": lat}, transpose=True)
+
+    with pytest.warns(FieldSetWarning):
+        # allow_time_extrapolation with time_periodic warning
+        fieldset = FieldSet.from_data(
+            data={"U": u, "V": v},
+            dimensions={"lon": lon, "lat": lat},
+            transpose=True,
+            allow_time_extrapolation=True,
+            time_periodic=1,
+        )
+
+    # b-grid with s-levels and POP output in meters
+    mesh = os.path.join(os.path.join(os.path.dirname(__file__), "test_data"), "POPtestdata_time.nc")
+    filenames = mesh
+    variables = {"U": "U", "V": "V", "W": "W", "T": "T"}
+    dimensions = {"lon": "lon", "lat": "lat", "depth": "w_deps", "time": "time"}
+    with pytest.warns(FieldSetWarning):
+        fieldset = FieldSet.from_pop(filenames, variables, dimensions, mesh="flat")
+
+    with pytest.warns(FieldSetWarning):
+        # timestamps with time in file warning
+        fieldset = FieldSet.from_pop(filenames, variables, dimensions, mesh="flat", timestamps=[0, 1, 2, 3])
+
+
+def test_kernel_warnings():
+    # positive scaling factor for W
+    mesh = os.path.join(os.path.join(os.path.dirname(__file__), "test_data"), "POPtestdata_time.nc")
+    filenames = mesh
+    variables = {"U": "U", "V": "V", "W": "W", "T": "T"}
+    dimensions = {"lon": "lon", "lat": "lat", "depth": "w_deps", "time": "time"}
+    fieldset = FieldSet.from_pop(filenames, variables, dimensions, mesh="flat")
+    fieldset.W._scaling_factor = 0.01
+    pset = ParticleSet(fieldset=fieldset, pclass=ScipyParticle, lon=[0], lat=[0], depth=[0], time=[0])
+    with pytest.warns(KernelWarning):
+        pset.execute(AdvectionRK4_3D, runtime=1, dt=1)
+
+    # RK45 warnings
+    lat = [0, 1, 5, 10]
+    lon = [0, 1, 5, 10]
+    u = [[1, 1, 1, 1] for _ in range(4)]
+    v = [[1, 1, 1, 1] for _ in range(4)]
+    fieldset = FieldSet.from_data(data={"U": u, "V": v}, dimensions={"lon": lon, "lat": lat}, transpose=True)
+    pset = ParticleSet(
+        fieldset=fieldset,
+        pclass=ScipyParticle.add_variable("next_dt", dtype=np.float32, initial=1),
+        lon=[0],
+        lat=[0],
+        depth=[0],
+        time=[0],
+        next_dt=1,
+    )
+    with pytest.warns(KernelWarning):
+        pset.execute(AdvectionRK45, runtime=1, dt=1)
